[POC][WIP] Use source plugin to inject always
This is a secondary step to add loading of the injection code into calls
that otherwise wouldn't get loaded by the `Bundler::Plugin.hook`
mechanics.

Since a portion of `Bundler::LockfileParser#parse_source` will trigger a
`Bundler::Plugin.source_from_lock` call, we can use that as a entry
point to prepend code into the `Bundler` internals to trigger reading
the other sources.

This does mean we need to override something that is already going to be
called after the parse, and `Bundler::Definition#validate_runtime!` was
chosen since it is called fairly frequently in most `Bundler::CLI` and
`bundler/setup` invocations.

The code for `inject!` was now split up to handle the two cases:

- Hooks
- Sources

But share what it can through the `.injection_builder` method.

One final thing this requires is a separate "gem" for the dummy "Source"
that we are using to make sure it is injected earlier, and an extra call
in the Gemfile to trigger it properly:

**Gemfile**

  plugin "bundler-inject", "~>0.1.0", :git => "https://github.com/Fryguy/bundler-inject"
  source("bundler-inject", :type => :bundle_inject ) {}

The `"bundler-inject"` string could be anything (currently), but the
main point is the `:type`, which gets converted to
`bundle-source_bundle_inject` in the `Bundler::Plugin::Dsl#source`
method.  The above could probably also be condensed down to just the
source line, if we wanted (if/when we release a plublic gem for this).

This is the one part I am confused about because the [specs][1] make it
seem like you shouldn't need to do this (have a extra gem) but for some
reason I haven't found a way to get this to work without putting the
extra `bundle-source_bundle_inject.gemspec` in place.

That said, if it does end up being required, we can always have it as a
gem stub anyway for versions of bundler that need it (where the plugin
system is still being worked on).  That said, we probably want to
propose improvements to the plugin system to allow hook points in the
future to avoid all of this craziness.

POC TODO:  AKA "the Still Not Working parts"
--------------------------------------------

There are still some issues when running `bundle check`, specifically
this still doesn't handle the lockfile properly, which probably means we
need to re-update the definition after doing the `eval_gemfile` calls.
But this mostly is a working prototype as the code for the plugin is
loaded prior to getting an error, which is the main point this is trying
to solve.

I will work on reverse engineering the lockfile validation and push to
this branch when I get that working, but that should be the last piece
required.

[1]: https://github.com/bundler/bundler/blob/5dc368b07c407702a485ed53d8fd8a8ab40fa9ef/spec/plugins/source_spec.rb#L61-L84